feat: Cabinet::all_files
Currently, there is no method to retrieve all files in the cabinet. Library users who want to extract all files need to iterate using Cabinet::read_file, which throws away bytes up to where the target file is located in the decompressed stream, which is inefficient.

This commit adds a new method Cabinet::all_files() that returns an AllFiles struct. When library user repeatedly calls next_file method on it, it yields tuples of (file_entry, reader) for each file in the cabinet.